mage_integrations/requirements.txt
google-analytics-data==0.14.2
google-api-python-client==2.65.0
google-cloud-bigquery~=2.34.4
jsonschema==2.6.0
mysql-connector-python==8.0.31
psycopg2-binary~=2.8
python-dateutil==2.8.2
redshift-connector==2.0.909
requests~=2.27.0
simplejson==3.11.1
singer-python==5.12.2
snowflake-connector-python==2.7.9
sshtunnel==0.4.0
stripe==4.2.0
tap-framework==0.1.1
xmltodict==0.13.0
